Rename integration tests `self` to `local`
