+++ tests/integration/suite/actors/lock/call/reentry/reentry.go
@@ -17,5 +17,5 @@ import (
 	_ "github.com/dapr/dapr/tests/integration/suite/actors/lock/call/reentry/max"
 	_ "github.com/dapr/dapr/tests/integration/suite/actors/lock/call/reentry/pertype"
 	_ "github.com/dapr/dapr/tests/integration/suite/actors/lock/call/reentry/remote"
-	_ "github.com/dapr/dapr/tests/integration/suite/actors/lock/call/reentry/self"
+	_ "github.com/dapr/dapr/tests/integration/suite/actors/lock/call/reentry/local"
 )
